+++ src/Typin/Typin.Core/ICliMode.cs
@@ -1,0 +1,16 @@
+﻿namespace Typin
+{
+    using System.Threading.Tasks;
+    using Typin.Internal;
+
+    /// <summary>
+    /// CLI mode definition.
+    /// </summary>
+    public interface ICliMode
+    {
+        /// <summary>
+        /// Executes CLI mode.
+        /// </summary>
+        ValueTask Execute(ICliCommandExecutor executor);
+    }
+}