+++ src/Typin/Typin.Core/Internal/Extensions/TypeExtensions.cs
@@ -8,9 +8,11 @@
 
     internal static class TypeExtensions
     {
-        public static bool Implements(this Type type, Type interfaceType)
+        public static Type? TryGetEnumerableArgumentUnderlyingType(this PropertyInfo? property)
         {
-            return type.GetInterfaces().Contains(interfaceType);
+            return property != null && property.PropertyType != typeof(string)
+                       ? property.PropertyType.TryGetEnumerableUnderlyingType()
+                       : null;
         }
 
         public static Type? TryGetNullableUnderlyingType(this Type type)
@@ -34,39 +36,6 @@
                        .Where(t => t != null)
                        .OrderByDescending(t => t != typeof(object)) // prioritize more specific types
                        .FirstOrDefault();
-        }
-
-        public static MethodInfo GetToStringMethod(this Type type)
-        {
-            return type.GetMethod(nameof(ToString), Type.EmptyTypes);
-        }
-
-        public static bool IsToStringOverriden(this Type type)
-        {
-            return type.GetToStringMethod() != typeof(object).GetToStringMethod();
-        }
-
-        public static MethodInfo GetStaticParseMethod(this Type type, bool withFormatProvider = false)
-        {
-            Type[] argumentTypes = withFormatProvider
-                ? new[] { typeof(string), typeof(IFormatProvider) }
-                : new[] { typeof(string) };
-
-            return type.GetMethod("Parse",
-                                  BindingFlags.Public | BindingFlags.Static,
-                                  null,
-                                  argumentTypes,
-                                  null);
-        }
-
-        public static Array ToNonGenericArray<T>(this IEnumerable<T> source, Type elementType)
-        {
-            ICollection sourceAsCollection = source as ICollection ?? source.ToArray();
-
-            Array array = Array.CreateInstance(elementType, sourceAsCollection.Count);
-            sourceAsCollection.CopyTo(array, 0);
-
-            return array;
         }
     }
 }