+++ src/Typin/Typin.Core/Internal/Polyfills.cs
@@ -5,71 +5,11 @@
 #if NETSTANDARD2_0
 namespace System
 {
-    using Linq;
-
     internal static class Extensions
     {
-        public static bool Contains(this string str, char c)
-        {
-            return str.Any(i => i == c);
-        }
-
         public static bool StartsWith(this string str, char c)
         {
             return str.Length > 0 && str[0] == c;
-        }
-
-        public static bool EndsWith(this string str, char c)
-        {
-            return str.Length > 0 && str[str.Length - 1] == c;
-        }
-
-        public static string[] Split(this string str, char separator, StringSplitOptions options = StringSplitOptions.None)
-        {
-            return str.Split(new char[] { separator }, options);
-        }
-    }
-}
-
-namespace System.Collections.Generic
-{
-    using System.Linq;
-
-    internal static class Extensions
-    {
-        public static void Deconstruct<TKey, TValue>(this KeyValuePair<TKey, TValue> pair, out TKey key, out TValue value)
-        {
-            key = pair.Key;
-            value = pair.Value;
-        }
-
-        public static TValue GetValueOrDefault<TKey, TValue>(this IReadOnlyDictionary<TKey, TValue> dic, TKey key)
-        {
-            return dic.TryGetValue(key!, out TValue result) ? result! : default!;
-        }
-
-        public static bool TryAdd<TKey, TValue>(this IDictionary<TKey, TValue> dic, TKey key, TValue value)
-        {
-            if (!dic.ContainsKey(key))
-            {
-                dic.Add(key, value);
-                return true;
-            }
-
-            return false;
-        }
-
-        public static bool TryGetValue<T>(this HashSet<T> hashSet, T equalValue, out T actualValue)
-        {
-            if (!hashSet.Contains(equalValue))
-            {
-                actualValue = hashSet.Where(x => x.Equals(equalValue)).Single();
-                return true;
-            }
-
-            actualValue = default;
-
-            return false;
         }
     }
 }